relacionamento n:n

+++ Controllers/MetasController.cs
@@ -12,6 +12,7 @@ namespace pomodoro.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+   
     public class MetasController : ControllerBase
     {
         private readonly ApiContext _context;
@@ -23,36 +24,45 @@ namespace pomodoro.Controllers
 
         // GET: api/Metas
         [HttpGet]
+        [Authorize]
         public async Task<ActionResult<IEnumerable<Meta>>> GetMetas()
         {
-          if (_context.Metas == null)
-          {
-              return NotFound();
-          }
-            return await _context.Metas.ToListAsync();
+            if (_context.Metas == null)
+            {
+                return NotFound();
+            }
+            var userLog = await _context?.Usuarios?.Where(x => x.Login == User.Identity.Name ).FirstAsync();
+            return  userLog.Metas.ToList();
         }
 
         // GET: api/Metas/5
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<Meta>> GetMeta(long id)
         {
           if (_context.Metas == null)
           {
               return NotFound();
           }
+
             var meta = await _context.Metas.FindAsync(id);
 
             if (meta == null)
             {
                 return NotFound();
             }
-
-            return meta;
+            var userLog = _context?.Usuarios?.Where(x => x.Login == User.Identity.Name ).FirstOrDefault();
+            if(userLog?.Metas.ToList().Exists(x => x.MetasId == meta.MetasId)??false){
+                return meta;
+            }else{
+                return Unauthorized();
+            }
         }
 
         // PUT: api/Metas/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> PutMeta(long id, Meta meta)
         {
             if (id != meta.MetasId)
@@ -64,7 +74,12 @@ namespace pomodoro.Controllers
 
             try
             {
-                await _context.SaveChangesAsync();
+                var userLog = _context?.Usuarios?.Where(x => x.Login == User.Identity.Name ).FirstOrDefault();
+                if(userLog?.Metas.ToList().Exists(x => x.MetasId == meta.MetasId)??false){
+                    await _context.SaveChangesAsync();
+                }else{
+                    return Unauthorized();
+                }
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -87,12 +102,25 @@ namespace pomodoro.Controllers
         [Authorize]
         public async Task<ActionResult<Meta>> PostMeta(Meta meta)
         {
-            if (_context.Metas == null)
+            if (_context?.Metas is null)
             {
                 return Problem("Entity set 'ApiContext.Metas'  is null.");
             }
+            if (_context?.Usuarios is null)
+            {
+                return Problem("Entity set 'ApiContext.Metas'  is null.");
+            }
+            if(User?.Identity is null){
+                return Unauthorized();
+            }
             var userLog = _context.Usuarios.Where(x => x.Login == User.Identity.Name ).FirstOrDefault();
+            if(userLog is null){
+                return Unauthorized();
+            }
             _context.Metas.Add(meta);
+            if(userLog.Metas is null){
+                userLog.Metas = new List<Meta>();
+            }
             userLog?.Metas.Add(meta);
             await _context.SaveChangesAsync();
 
@@ -103,6 +131,7 @@ namespace pomodoro.Controllers
 
         // DELETE: api/Metas/5
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteMeta(long id)
         {
             if (_context.Metas == null)
@@ -114,11 +143,15 @@ namespace pomodoro.Controllers
             {
                 return NotFound();
             }
+            var userLog = _context?.Usuarios?.Where(x => x.Login == User.Identity.Name ).FirstOrDefault();
+            if(userLog?.Metas.ToList().Exists(x => x.MetasId == meta.MetasId)??false){
+                _context.Metas.Remove(meta);
+                await _context.SaveChangesAsync();
 
-            _context.Metas.Remove(meta);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
+                return NoContent();
+            }else{
+                return Unauthorized();
+            }
         }
 
         private bool MetaExists(long id)
